Remove Int - porting Decimal to Int is WIP
